+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PASSBOLT-1566: Share autocomplete html is not valid
 - PASSBOLT-1778: Simplify toolbarController openPassboltTab function
 - PASSBOLT-1680: Password is limited to 50 chars, increase the limit to 4096
+- PASSBOLT-1657: As AP I should not be able to complete the recovery process with my public key
 
 ## [1.3.0]
 ### Added